fix(数据服务): 修复分页问题并改进版本选择逻辑
- 更新`DataService.InspectVersionsAsync`方法以正确处理分页，确保所有版本都被获取。
- 向`LazyObject`添加回调以在获取值后执行操作。
- 更新`ExhibitPackageModal`以使用新的回调逻辑来设置所选版本。
- 更新`PackageEntryModal`以使用新的回调逻辑来设置所选版本。
- 更新`CurseForgeRepository`以正确处理分页。
- 更新`LoaderEditorDialog`以使用新的`LazyObject`构造函数。
- 更新`ExhibitPackageButton`图标。
- 添加`ExhibitPackageModal`的命令。
- 更新`ExhibitPackageModal`的UI以支持不同的展览状态。
- 更新`PackageExplorerViewModel`以使用`ModifyPending`回调。
- 更新`LazyContainer`以处理`LazyObject`中的空值。
- 更新`PackageEntryModal`的UI以包含一个定位按钮。

+++ src/Polymerium.Trident/Services/CurseForgeService.cs
@@ -235,11 +235,11 @@ public class CurseForgeService(ICurseForgeClient client)
         uint modId,
         string? gameVersion,
         ModLoaderTypeModel? modLoader,
-        int index = 0,
-        int count = 50)
+        uint index = 0,
+        uint count = 50)
     {
         var rv = await client
-                      .GetModFilesAsync(modId, gameVersion, modLoader, (uint)index, (uint)count)
+                      .GetModFilesAsync(modId, gameVersion, modLoader, index, count)
                       .ConfigureAwait(false);
         return rv;
     }
